requirements.txt
flask-socketio==2.9.3
eventlet==0.21.0
Flask==0.12.2
rpi.gpio==0.6.3